split processor.c in two files
